Initial checkin of rtty-rpi project
This code originally started as an audio FSK generator for teletype
signaling. With some simple circuits, using the RPI GPIO pins eliminates
the need to have an AFSK demodulator, and directly drive a Teletype
using a Raspberry Pi. This code is the refactoring of the original
project, which used code from an DTMF tone generator, to create the
audio signaling.